fix: P0 batch 2 — TextRecord X/Y offsets, lossless bitmap format
- TextRecord stores OffsetX/OffsetY independently (was collapsed into one Point,
  so a record with only one axis re-encoded with both flags and an extra field).
- DefineBitsLossless(2): unknown bitmap format -> SwfFormatException; ColorMap8
  NumColors now the true color count (raw byte + 1, spec's "count - 1" field),
  encoded back as count - 1 so files still round-trip byte-for-byte.

+++ src/ShockwaveFlash/Types/Bitmap/BitmapFormat.cs
@@ -2,13 +2,13 @@ namespace ShockwaveFlash.Types.Bitmap;
 
 public abstract record BitmapFormat
 {
-    public sealed record BitmapFormatColorMap8(byte NumColors) : BitmapFormat;
+    public sealed record BitmapFormatColorMap8(int NumColors) : BitmapFormat;
 
     public sealed record BitmapFormatRgb15 : BitmapFormat;
 
     public sealed record BitmapFormatRgb32 : BitmapFormat;
 
-    public static BitmapFormat ColorMap8(byte numColors)
+    public static BitmapFormat ColorMap8(int numColors)
     {
         return new BitmapFormatColorMap8(numColors);
     }
